#22 Rename submodule and data class (zarr → ms)
