+++ drs4/specs/__init__.py
@@ -1,8 +1,8 @@
-__all__ = ["common", "csv", "vdif", "zarr"]
+__all__ = ["common", "csv", "ms", "vdif"]
 
 
 # submodules
 from . import common
 from . import csv
+from . import ms
 from . import vdif
-from . import zarr
